done homework 6
